Refactored configs

+++ scripts/git.sh
@@ -18,13 +18,16 @@ cd $REPOSITORY_FOLDER
 echo Git project folders setup. 
 while :
 do
-  echo Enter projects group folder name. Enter empty string to finish
+  echo "Enter projects group name (will be used in config files). Enter empty string to finish"
+  read PROJECT_GROUP_NAME
+  if [[ $PROJECT_GROUP_NAME == '' ]]; then break; fi
+  echo "Enter projects group folder name. Enter empty string to finish"
   read PROJECT_GROUP_FOLDER_NAME
   if [[ $PROJECT_GROUP_FOLDER_NAME == '' ]]; then break; fi
   mkdir $PROJECT_GROUP_FOLDER_NAME
 
   mkdir -p ~/.config/git
-  GIT_CONFIG_FILE_PATH=~/.config/git/${FILE_POSTFIX}
+  GIT_CONFIG_FILE_PATH=~/.config/git/${PROJECT_GROUP_NAME}
   rm -f GIT_CONFIG_FILE_PATH
   touch GIT_CONFIG_FILE_PATH
